requirements.txt
huggingface-hub==0.10.1
ipykernel==6.15.2
jupyter==1.0.0
matplotlib==3.5.3
numpy==1.23.4
opencv-python==4.6.0.66
pandas==1.5.1
pathlib==1.0.1
Pillow==9.2.0
torch==1.13.0
requests==2.28.1
seaborn==0.12.0
transformers==4.24.0
scikit-learn==1.2.0
torchmetrics==0.11.0
datasets==2.8.0
jiwer==2.5.1
evaluate==0.4.0


